Add checkboxes
