+++ frontend/src/metabase/querying/components/FilterModal/FilterValuePicker/InlineValuePicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./InlineValuePicker";
