+++ frontend/src/metabase/querying/components/FilterModal/FilterValuePicker/SelectValuePicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./SelectValuePicker";
